Make groups service to divide view and logic
