+++ app/src/main/java/ru/vsu/csf/bakebudget/services/AuthService.kt
@@ -1,0 +1,2 @@
+package ru.vsu.csf.bakebudget.services
+
